Consolidate packages a bit
