+++ agent/agent-tooling/src/test/java/com/microsoft/applicationinsights/agent/internal/perfcounter/JmxDataFetcherTest.java
@@ -19,7 +19,7 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.perfcounter.jmx;
+package com.microsoft.applicationinsights.agent.internal.perfcounter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
